Move Fortran functions to a module
